refactor(daemon): split the wire protocol out of daemon-core
`daemon-core` was published, so its row in the split table read *published*
and looked settled. Published is not the same as being the right size to
depend on. Serve's imports of it:

  daemon.protocol   46
  daemon.bta         2
  daemon.devices     2
  daemon (root)      1

Forty-six of fifty-one are wire shapes, and to name one of them an extracted
preview server took the whole daemon with it — the JSON-RPC server, the
APNG/GIF/ffmpeg encoders, the sandbox lifecycle, incremental discovery, the
recording test generator, the XR session registry, the history archive. 653
public declarations to reach the ones that describe a request.

`:daemon-protocol` splits on shape versus behaviour, not client versus
server. Anything that only describes what crosses the wire moved; anything
that reads a file, opens a socket or computes a result stayed.
`HistoryDataDelta` names the rule: the delta shape moved, while
`HistoryDataDiff` — which reads two archived entries off disk to produce one
— did not, and neither did the internal mirror types it decodes through.

The package is unchanged and `:daemon:core` exposes the new module as `api`,
so no consumer needed an import change. Two things did need one, both from
the same-package false negative that keeps surfacing in this work: a smart
cast on `params.maxFps` in JsonRpcServer, which stops applying once the
property is public API in another module, and one test that had been reading
`HistoryDataDiffProduct` without an import.

ABI: daemon-core 7,317 lines -> 2,479, daemon-protocol 4,843. What is left
of serve's coupling to daemon-core is now legible as exactly five imports.

Registered in all four places and verified with the probe end to end: 17
contracts resolved, 0 leaks.
