+++ scripts/serve-seam-allowlist.json
@@ -87,7 +87,7 @@
     "ee.schimke.composeai.daemon.bta": "daemon-core",
     "ee.schimke.composeai.daemon.client": "daemon-client",
     "ee.schimke.composeai.daemon.devices": "daemon-core",
-    "ee.schimke.composeai.daemon.protocol": "daemon-core",
+    "ee.schimke.composeai.daemon.protocol": "daemon-protocol",
     "ee.schimke.composeai.data.layoutinspector": "data-layoutinspector-core",
     "ee.schimke.composeai.data.overrides": "data-preview-overrides-core",
     "ee.schimke.composeai.data.pseudolocale": "data-pseudolocale-core",
